Remove unused modules in `vm_latest`
