+++ core/lib/multivm/src/versions/vm_latest/tests/call_tracer.rs
@@ -3,17 +3,14 @@ use std::sync::Arc;
 use once_cell::sync::OnceCell;
 use zksync_types::{Address, Execute};
 
+use super::TestedLatestVm;
 use crate::{
     interface::{TxExecutionMode, VmExecutionMode, VmInterface},
     tracers::CallTracer,
-    vm_latest::{
-        constants::BATCH_COMPUTATIONAL_GAS_LIMIT,
-        tests::{
-            tester::VmTesterBuilder,
-            utils::{read_max_depth_contract, read_test_contract},
-        },
-        HistoryEnabled, ToTracerPointer,
+    versions::testonly::{
+        read_max_depth_contract, read_test_contract, ContractToDeploy, VmTesterBuilder,
     },
+    vm_latest::{constants::BATCH_COMPUTATIONAL_GAS_LIMIT, ToTracerPointer},
 };
 
 // This test is ultra slow, so it's ignored by default.
@@ -22,14 +19,14 @@ use crate::{
 fn test_max_depth() {
     let contarct = read_max_depth_contract();
     let address = Address::random();
-    let mut vm = VmTesterBuilder::new(HistoryEnabled)
+    let mut vm = VmTesterBuilder::new()
         .with_empty_in_memory_storage()
         .with_random_rich_accounts(1)
         .with_deployer()
         .with_bootloader_gas_limit(BATCH_COMPUTATIONAL_GAS_LIMIT)
         .with_execution_mode(TxExecutionMode::VerifyExecute)
-        .with_custom_contracts(vec![(contarct, address, true)])
-        .build();
+        .with_custom_contracts(vec![ContractToDeploy::account(contarct, address)])
+        .build::<TestedLatestVm>();
 
     let account = &mut vm.rich_accounts[0];
     let tx = account.get_l2_tx_for_execute(
@@ -54,16 +51,16 @@ fn test_max_depth() {
 
 #[test]
 fn test_basic_behavior() {
-    let contarct = read_test_contract();
+    let contract = read_test_contract();
     let address = Address::random();
-    let mut vm = VmTesterBuilder::new(HistoryEnabled)
+    let mut vm = VmTesterBuilder::new()
         .with_empty_in_memory_storage()
         .with_random_rich_accounts(1)
         .with_deployer()
         .with_bootloader_gas_limit(BATCH_COMPUTATIONAL_GAS_LIMIT)
         .with_execution_mode(TxExecutionMode::VerifyExecute)
-        .with_custom_contracts(vec![(contarct, address, true)])
-        .build();
+        .with_custom_contracts(vec![ContractToDeploy::account(contract, address)])
+        .build::<TestedLatestVm>();
 
     let increment_by_6_calldata =
         "7cf5dab00000000000000000000000000000000000000000000000000000000000000006";
