+++ core/lib/multivm/src/versions/vm_latest/tests/mod.rs
@@ -1,22 +1,27 @@
-use std::collections::HashSet;
+use std::collections::{HashMap, HashSet};
 
 use zk_evm_1_5_0::{
     aux_structures::{MemoryPage, Timestamp},
+    vm_state::VmLocalState,
     zkevm_opcode_defs::{ContractCodeSha256Format, VersionedHashLen32},
 };
-use zksync_types::{writes::StateDiffRecord, StorageKey, Transaction, H256, U256};
+use zksync_types::{writes::StateDiffRecord, StorageKey, StorageValue, Transaction, H256, U256};
 use zksync_utils::{bytecode::hash_bytecode, bytes_to_be_words, h256_to_u256};
-use zksync_vm_interface::{
-    CurrentExecutionState, L2BlockEnv, VmExecutionMode, VmExecutionResultAndLogs,
-};
 
 use super::{HistoryEnabled, Vm};
 use crate::{
-    interface::storage::{InMemoryStorage, StorageView},
+    interface::{
+        storage::{InMemoryStorage, ReadStorage, StorageView, WriteStorage},
+        CurrentExecutionState, L2BlockEnv, VmExecutionMode, VmExecutionResultAndLogs,
+    },
     versions::testonly::TestedVm,
     vm_latest::{
-        constants::BOOTLOADER_HEAP_PAGE, tests::tester::VmInstanceInnerState,
-        tracers::PubdataTracer, types::internals::TransactionData, TracerDispatcher,
+        constants::BOOTLOADER_HEAP_PAGE,
+        old_vm::{event_sink::InMemoryEventSink, history_recorder::HistoryRecorder},
+        tracers::PubdataTracer,
+        types::internals::TransactionData,
+        utils::logs::StorageLogQuery,
+        AppDataFrameManagerWithHistory, HistoryMode, SimpleMemory, TracerDispatcher,
     },
 };
 
@@ -46,13 +51,13 @@ mod rollbacks;
 mod secp256r1;
 mod simple_execution;
 mod storage;
-mod tester;
 mod tracing_execution_error;
 mod transfer;
 mod upgrade;
-mod utils;
 
-impl TestedVm for Vm<StorageView<InMemoryStorage>, HistoryEnabled> {
+type TestedLatestVm = Vm<StorageView<InMemoryStorage>, HistoryEnabled>;
+
+impl TestedVm for TestedLatestVm {
     type StateDump = VmInstanceInnerState<HistoryEnabled>;
 
     fn dump_state(&self) -> Self::StateDump {
@@ -160,5 +165,121 @@ impl TestedVm for Vm<StorageView<InMemoryStorage>, HistoryEnabled> {
         let tx = TransactionData::new(tx, false);
         let overhead = tx.overhead_gas();
         self.push_raw_transaction(tx, overhead, refund, true)
+    }
+}
+
+#[derive(Clone, Debug)]
+pub(crate) struct ModifiedKeysMap(HashMap<StorageKey, StorageValue>);
+
+impl ModifiedKeysMap {
+    fn new<S: ReadStorage>(storage: &mut StorageView<S>) -> Self {
+        let mut modified_keys = storage.modified_storage_keys().clone();
+        let inner = storage.inner_mut();
+        // Remove modified keys that were set to the same value (e.g., due to a rollback).
+        modified_keys.retain(|key, value| inner.read_value(key) != *value);
+        Self(modified_keys)
+    }
+}
+
+// We consider hashmaps to be equal even if there is a key
+// that is not present in one but has zero value in another.
+impl PartialEq for ModifiedKeysMap {
+    fn eq(&self, other: &Self) -> bool {
+        for (key, value) in &self.0 {
+            if *value != other.0.get(key).copied().unwrap_or_default() {
+                return false;
+            }
+        }
+        for (key, value) in &other.0 {
+            if *value != self.0.get(key).copied().unwrap_or_default() {
+                return false;
+            }
+        }
+        true
+    }
+}
+
+#[derive(Clone, PartialEq, Debug)]
+pub(crate) struct DecommitterTestInnerState<H: HistoryMode> {
+    /// There is no way to "truly" compare the storage pointer,
+    /// so we just compare the modified keys. This is reasonable enough.
+    pub(crate) modified_storage_keys: ModifiedKeysMap,
+    pub(crate) known_bytecodes: HistoryRecorder<HashMap<U256, Vec<U256>>, H>,
+    pub(crate) decommitted_code_hashes: HistoryRecorder<HashMap<U256, Option<u32>>, HistoryEnabled>,
+}
+
+#[derive(Clone, PartialEq, Debug)]
+pub(crate) struct StorageOracleInnerState<H: HistoryMode> {
+    /// There is no way to "truly" compare the storage pointer,
+    /// so we just compare the modified keys. This is reasonable enough.
+    pub(crate) modified_storage_keys: ModifiedKeysMap,
+    pub(crate) frames_stack: AppDataFrameManagerWithHistory<Box<StorageLogQuery>, H>,
+    pub(crate) paid_changes: HistoryRecorder<HashMap<StorageKey, u32>, H>,
+    pub(crate) initial_values: HistoryRecorder<HashMap<StorageKey, U256>, H>,
+    pub(crate) returned_io_refunds: HistoryRecorder<Vec<u32>, H>,
+    pub(crate) returned_pubdata_costs: HistoryRecorder<Vec<i32>, H>,
+}
+
+#[derive(Clone, PartialEq, Debug)]
+pub(crate) struct PrecompileProcessorTestInnerState<H: HistoryMode> {
+    pub(crate) timestamp_history: HistoryRecorder<Vec<Timestamp>, H>,
+}
+
+/// A struct that encapsulates the state of the VM's oracles
+/// The state is to be used in tests.
+#[derive(Clone, PartialEq, Debug)]
+pub(crate) struct VmInstanceInnerState<H: HistoryMode> {
+    event_sink: InMemoryEventSink<H>,
+    precompile_processor_state: PrecompileProcessorTestInnerState<H>,
+    memory: SimpleMemory<H>,
+    decommitter_state: DecommitterTestInnerState<H>,
+    storage_oracle_state: StorageOracleInnerState<H>,
+    local_state: VmLocalState,
+}
+
+impl<S: ReadStorage, H: crate::glue::history_mode::HistoryMode> Vm<StorageView<S>, H> {
+    // Dump inner state of the VM.
+    pub(crate) fn dump_inner_state(&self) -> VmInstanceInnerState<H::Vm1_5_0> {
+        let event_sink = self.state.event_sink.clone();
+        let precompile_processor_state = PrecompileProcessorTestInnerState {
+            timestamp_history: self.state.precompiles_processor.timestamp_history.clone(),
+        };
+        let memory = self.state.memory.clone();
+        let decommitter_state = DecommitterTestInnerState {
+            modified_storage_keys: ModifiedKeysMap::new(
+                &mut self
+                    .state
+                    .decommittment_processor
+                    .get_storage()
+                    .borrow_mut(),
+            ),
+            known_bytecodes: self.state.decommittment_processor.known_bytecodes.clone(),
+            decommitted_code_hashes: self
+                .state
+                .decommittment_processor
+                .get_decommitted_code_hashes_with_history()
+                .clone(),
+        };
+
+        let storage_oracle_state = StorageOracleInnerState {
+            modified_storage_keys: ModifiedKeysMap::new(
+                &mut self.state.storage.storage.get_ptr().borrow_mut(),
+            ),
+            frames_stack: self.state.storage.storage_frames_stack.clone(),
+            paid_changes: self.state.storage.paid_changes.clone(),
+            initial_values: self.state.storage.initial_values.clone(),
+            returned_io_refunds: self.state.storage.returned_io_refunds.clone(),
+            returned_pubdata_costs: self.state.storage.returned_pubdata_costs.clone(),
+        };
+        let local_state = self.state.local_state.clone();
+
+        VmInstanceInnerState {
+            event_sink,
+            precompile_processor_state,
+            memory,
+            decommitter_state,
+            storage_oracle_state,
+            local_state,
+        }
     }
 }
